requirements.txt
scipy==1.4.1
gym==0.17.1
torch==1.0.0
pandas==0.22.0
numpy==1.15.4
matplotlib==3.2.1
GitPython==3.1.3
scikit_learn==0.23.1
typing==3.7.4.1
